src/test/java/com/wizeek/leetcode/Solution496Test.java
package com.wizeek.leetcode;

import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertArrayEquals;

public class Solution496Test {

    private Solution496 solution;

    @Before
    public void setUp() {
        solution = new Solution496();
    }

    @Test
    public void test1() {
        assertArrayEquals(new int[]{-1, 3, -1}, solution.nextGreaterElement(new int[]{4, 1, 2}, new int[]{1, 3, 4, 2}));
    }

    @Test
    public void test2() {
        assertArrayEquals(new int[]{3, -1}, solution.nextGreaterElement(new int[]{2, 4}, new int[]{1, 2, 3, 4}));
    }
}
